// reactjs-site/src/types/content.type.ts
export interface FeatureItem {
  icon: string;
  titleKey: string;
  descKey: string;
  tagKey: string;
}

export interface ProblemItem {
  icon: string;
  titleKey: string;
  descKey: string;
}

export interface StepItem {
  titleKey: string;
  descKey: string;
}

export interface TestimonialItem {
  stars: string;
  textKey: string;
  avatar: string;
  name: string;
  roleKey: string;
}

export interface MetricItem {
  value: string;
  labelKey: string;
}

export interface PlatformItem {
  icon: string;
  label: string;
}

export interface TrustItem {
  textKey: string;
}